SOUNDS! Created a sound system from scratch. Now sounds can be played by creating a new instance of Sound() choosing the file path of the sound and writing sound.play(). Sounds also have panning and volume features, I also created a method where the sounds come out from different speakers depending on the play location (NOTE: This only works on some systems like Windows, if the system is not supported, it will simply try a different approaching by controlling the volume). I also fixed some collision issues and added a new method for custom image sizes and level backgrounds (setting the level background seems to lag so the method responsible for it, is commented and disabled).
EDIT: Added projectiles and remade the animation class. The old animation class is now named Animation_DELETE_PENDING, the class was not deleted in case there are any references to it in the project. As soon as the references are removed and the animations migrate to the new class, it'll be deleted, for now, it stays as a deprecated class.

Current Issues:
Since the object updates run in parallel, the arrayList tends to crash due to performing various operations at the same time <<NEEDS FIXING ASAP